fix(intel): validate non-secure addresses and enforce response length in mailbox/SMC handlers
The SiP v2 mailbox path and several v3 SMC handlers passed
non-secure-world addresses straight to EL3 mailbox operations without
checking they point into the non-secure DDR range, letting a
normal-world caller read or write secure memory. The v3 response
parser also copied as many words as the SDM returned, ignoring the
caller's buffer length.

  - Add is_address_in_ddr_range() checks to the v2 send/poll path and
    to the v3 device identity, QSPI, generic mailbox, FCS
    random/provision and MCTP handlers before any access.
  - Bound the v3 response copy to the caller-declared cb_args_len.
  - Widen cb_args_len to uint32_t so it is no longer truncated, and
    plumb the response buffer length through to the parser.
  - Reject a v2 poll data size smaller than one word so the length
    computation cannot underflow.

+++ plat/intel/soc/common/include/socfpga_mailbox.h
@@ -371,7 +371,7 @@ typedef struct sdm_command {
 	uint32_t flags;
 	sdm_command_callback cb;
 	uint32_t *cb_args;
-	uint8_t cb_args_len;
+	uint32_t cb_args_len;
 } sdm_command_t;
 
 /* Get the transcation ID from client ID and job ID. */
